server/requirements.txt
django==5.2
djangorestframework==3.16.0
django-filter==25.1
Pillow==11.2.1
django-cors-headers==4.7.0
django_q2==1.7.6
weasyprint==65.1
libsass==0.23.0
markdown==3.8
pygments==2.19.1
drf-spectacular==0.28.0
cvss==3.4
jinja2==3.1.6
argon2-cffi==23.1.0
rest-framework-generic-relations==2.2.0
nh3==0.2.21
defusedxml==0.8.0rc2
matplotlib==3.10.1
djangoql==0.18.1
tldextract==5.3.0
psycopg2-binary==2.9.10
django-vite==3.1.0
django-allauth==65.7.0
django-allauth[socialaccount]==65.7.0
croniter==6.0.0